Protect against link in link with summaries
There was a warning in the notes that links don't work in summaries, but it is easy, especially when bringing in old content, to have a first paragraph that has a link. And, in any event, this should be protected from by the template rather than a potential source of user error that breaks the theme.

We use .Summary | htmlUnescape | plainify to resolve this (and likewise for .Description).

Signed-off-by: Daniel F. Dickinson <daniel-f-dickinson@outlook.com>