Protect against link in link with summaries (#138)
* Protect against link in link with summaries

There was a warning in the notes that links don't work in summaries, but it is easy, especially when bringing in old content, to have a first paragraph that has a link. And, in any event, this should be protected from by the template rather than a potential source of user error that breaks the theme.

We use .Summary | htmlUnescape | plainify to resolve this (and likewise for .Description).

Signed-off-by: Daniel F. Dickinson <daniel-f-dickinson@outlook.com>

* Update summary.html

Good concept. The one thing I need to change: A string with Markdown markup and HTML entities needs to get rendered first and then stripped of the resulting HTML markup. That’s why I’m changing the order of the pipes. 

Signed-off-by: Georg Makowski <gm@desider.at>

---------

Signed-off-by: Daniel F. Dickinson <daniel-f-dickinson@outlook.com>
Signed-off-by: Georg Makowski <gm@desider.at>
Co-authored-by: Georg Makowski <gm@desider.at>